Move decorator to submodule, import needed imports
